Testing, Summary, and words
Finished off adding the 'words' property to the grid model and
calculating the current set of words based on the current grid.

Added unit tests for the word caluclation utility functions and as a
result pulled in a number of other modules into the development
environment.

Exploited the new word properties in to some basic reporting on the
screen of the general statistics of the puzzle as well as the start of
a cluing section.

Added some header buttons that don't currently do anything but will
add some functionality in the future.