Reducer/Context refactoring
Refactored to use a proper useContext and reducer functionality for
the crossword state.

Changed the primary state object from grid to be the crossword as a
whole.

Refactored the command class strategy into functional components.

Reduced the complexity of the presentation components, off loading
most of the logic into the context and reducer. Still significant
amounts of duplicate work going on in the reducer/context that will
probably need to be refactored when unit/functional tests are up and
running which should probably be the next step.

One casualty in the refactor is word lists. The current word highlight
mechanic is working fine, but part of the refactor was to start to use
ViewModels for the grid and the word list since the core model is the
same.