Word View synced to grid
The word list view will now automatically filter to the current word
an orientation providing a list of words based on the current contents
as well as length. This required another custom hook called
useCurrentWord. Thinking this needs to be refactored with the wordlist
hook and some of the other grid utility functions as there is
duplicate calculations being made that will need to be cleaned up.