New Crossword
Added functionality to the crossword reducer to create a new crossword
with custom sizing. Added a basic modal to fire off the request for a
new crossword. Will probably need to clean up some of the language and
possible move this up the ladder to the application context reducer
but for now it will live with the crossword reducer.