Undoing
Added undo functionality including a commandStack to the crossword
context.

Radically simplified the reducer/context operation. The reducer
payload is not the command itself which is run through the dispatch
function. There are now only 3 actions to the dispatch reducer:

CrosswordCommand,
UndoCrosswordCommand,
NewCrossword Command.

NewCrossword command is probably overkill and maybe removed in the
future since it is the exact same as the standard crossword action.

A lot of functionlity was moved from the reducer into either the
commands themselves or the context function.

The context could probably stand for a refactor as well (that if/else block
looks ugly as sin). Maybe convert to a dictionary of functions.