Grid refactoring
Instead of maintaining the grid as a matrix within the main crossword
context, it will now be a flat array. Partially this is because the
.puz file is structured this way, with the actual grid shape being
specified via the width/height values but also because it makes
certain functions easier.

As a result the main crossword context also provides a number of
'views' that transform the flat array into a variety of alternative
datasets.

This includes:

GridView: A matrix representation of the flat array
WordView: A list of words with indicies, orientation, and values
CurrentWord: The current word

The result is a some what cleaner set of utility functions and a
single class that calls them which should reduce impacts of future
changes so long as the function signatures don't have to change. This
will also set up a better framework for unit testing which I'm going
to need to re-write.

The other major change is that the command dispatch function now takes
as a payload a list of commands to be fired in order (curried). This
makes the dispatcher a bit more flexible as multiple commands can be
dispatched at the same time instead of worring about rendered order of
multiple dispatches firing.