Icon updates
In the previous commit I added the react-icons package to handle the
icons for the main menu. This was nice and convient but a gigantic
library for only a hand full of icons I needed. As a result it was
playing merry hell with LSP while deving. I've removed the package and
instead pulled the icons in as svgs into the assets directory as
reference and created react components for each of the icons that I
need. It isn't a great solution as I will now need to create a new
react component for any other icons I need but node is no longer
crying in pain while I'm trying to build/develop which is a plus.