Details, Details
- Replaced the SummaryLayout with the DetailsLayout as it is more
accurate. Wired the text input fields to the crossword context and
added notes and copyright into the mix.
- Added standard ui components for text inputs and text areas along with some basic
styling. Added updateMetadataCommands to the mix to update the
context. Currently they also set the position to -1 because otherwise
very update would shift the focus back to the grid. This is due to the
crossword context updates forcing a rerender of the grid even if the
grid itself in the context doesn't change. This bears some
consideration on whether or not that should be refactored into a
seperate context/nested context since re-rendering the grid on
metadata updates is not necessary and is just wasted cpu cycles.