First unit test
First crack at unit testing. Starting small with the square input
component. This made me want to refactor how it was taking in state
from the gridlayout. Which led to a lot of cascading changes and
breakages. Stashing it in this new unit testing branch while I work on
it.