Header Sidebar functional tests
Added functional tests for the header sidebar interactions as their
behavior and interactions have been vexing me. Added functional tests
for using the button to open and close the menu as well as closing by
losing focus on the menu. Added some accessibility tests to the
sidebar widget itself. There are probably more accessibility functions
that need to be added but otherwise I think its in pretty good shape.