Keyboard Header updates
Added the ability to tab into the header and toggle it via the
keyboard.

Added functional tests for toggle button keyboard toggling as well as
integration tests for the sidebar and header. Visual styling for the
header focus needs to be updated a bit but it is working for the most part.