Tab refactor and re-org
Updated all Tab Views to be layouts and to use the tab panel container
component. Updated and refactored how tab definitions function. Still
not in love with how the tabs are being displayed but will leave for
another commit to resolve as this one is already getting quite
large. Stories to come next for the various view -> layout conversions.